requirements.txt
# Django
django>=3.1,<3.2

# DRF
djangorestframework>=3.12,<3.13